Updates to bus-plan-run for debug and skipping solver

+++ bus-plan-run.py
@@ -7,6 +7,7 @@ import signal
 class BusPlanConfig():
     def __init__(self,
                  code_dir,
+                 run_solver,
                  cost_matrix_csv,
                  student_dataset_csv,
                  student_stop_eligibility_csv,
@@ -19,6 +20,7 @@ class BusPlanConfig():
                  router_output_csv,
                  analysis_output_student_ridetimes_csv):
         self.code_dir = code_dir
+        self.run_solver = run_solver
         self.cost_matrix_csv = cost_matrix_csv
         self.student_dataset_csv = student_dataset_csv
         self.student_stop_eligibility_csv = student_stop_eligibility_csv
@@ -43,18 +45,51 @@ def get_student_stop_eligibility_csv(scenario):
 def call_sbt(args):
     call('sbt "{}"'.format(' '.join(args)), shell=True)
 
+def fix_up_solver_routes(solver_output_routes_csv):
+    """
+    TODO: REMOVE AFTER https://github.com/azavea/bus-plan/issues/25 RESOLUTION
+
+    Fixes up the solver output to remove any empty stops that happen after a school
+    """
+    from functools import reduce
+
+    def fold_func(acc, part):
+        (new_line, school_found) = acc
+        if school_found:
+            return (new_line, school_found)
+        else:
+            new_line.append(part.strip())
+            return (new_line, part.startswith("school"))
+
+    print("FIXING UP {}".format(solver_output_routes_csv))
+
+    new_lines = []
+    with open(solver_output_routes_csv) as f:
+        for line in f.readlines():
+            (line_parts, _) = reduce(fold_func, line.split(','), ([], False))
+            new_line = ','.join(line_parts)
+            new_lines.append(new_line)
+    with open(solver_output_routes_csv, 'w') as f:
+        f.write('\n'.join(new_lines))
+
+
 def run_config(config):
     os.chdir(config.code_dir)
 
     # Run the solver
-    call_sbt(["planner/runMain",
-              "com.azavea.BusPlanner",
-              config.cost_matrix_csv,
-              config.student_dataset_csv,
-              config.student_stop_eligibility_csv,
-              config.garage_count_csv,
-              config.solver_output_routes_csv,
-              config.solver_output_student_assignment_csv])
+    if config.run_solver:
+        call_sbt(["planner/runMain",
+                  "com.azavea.BusPlanner",
+                  config.cost_matrix_csv,
+                  config.student_dataset_csv,
+                  config.student_stop_eligibility_csv,
+                  config.garage_count_csv,
+                  config.solver_output_routes_csv,
+                  config.solver_output_student_assignment_csv])
+
+    #### TODO: REMOVE AFTER https://github.com/azavea/bus-plan/issues/25 RESOLUTION ####
+    # Fix up the solver output to remove any empty stops that happen after a school
+    fix_up_solver_routes(config.solver_output_routes_csv)
 
     # Run the router
     call_sbt(["otp/runMain",
@@ -73,7 +108,7 @@ def run_config(config):
           config.solver_output_student_assignment_csv,
           config.analysis_output_student_ridetimes_csv])
 
-def run(data_dir, output_dir, code_dir):
+def run(data_dir, output_dir, code_dir, run_solver, debug_run):
     def d(p):
         return os.path.join(data_dir, p)
 
@@ -89,11 +124,18 @@ def run(data_dir, output_dir, code_dir):
     graph_without_students_obj = d("graph_withoutStudents.obj")
 
     confs = []
-    for scenario in list(scenarios.keys())[:1]:
+    scenario_names = scenarios.keys()
+    if debug_run:
+        scenario_names = list(scenario_names)[:1]
+    for scenario in scenario_names:
         student_stop_eligibility_csv = d(get_student_stop_eligibility_csv(scenario))
 
         # Do 4 runs of each scenario
-        for x in range(0, 2):
+        r = range(0, 2)
+        if debug_run:
+            r = range(0, 4)
+
+        for x in r:
             run_output_dir = os.path.join(output_dir, "{}_run{}".format(scenario, x))
             if not os.path.isdir(run_output_dir):
                 os.makedirs(run_output_dir)
@@ -108,6 +150,7 @@ def run(data_dir, output_dir, code_dir):
 
 
             confs.append(BusPlanConfig(code_dir,
+                                       run_solver,
                                        cost_matrix_csv,
                                        student_dataset_csv,
                                        student_stop_eligibility_csv,
@@ -120,25 +163,38 @@ def run(data_dir, output_dir, code_dir):
                                        router_output_csv,
                                        analysis_output_student_ridetimes_csv))
 
-    # Deal with Ctrl-C weirdness
-    # https://stackoverflow.com/questions/11312525/catch-ctrlc-sigint-and-exit-multiprocesses-gracefully-in-python
-    original_sigint_handler = signal.signal(signal.SIGINT, signal.SIG_IGN)
-    pool = Pool(processes=len(confs))
-    signal.signal(signal.SIGINT, original_sigint_handler)
+    if not debug_run:
+        # Deal with Ctrl-C weirdness
+        # https://stackoverflow.com/questions/11312525/catch-ctrlc-sigint-and-exit-multiprocesses-gracefully-in-python
+        original_sigint_handler = signal.signal(signal.SIGINT, signal.SIG_IGN)
+        pool = Pool(processes=len(confs))
+        signal.signal(signal.SIGINT, original_sigint_handler)
 
-    try:
-        res = pool.map_async(run_config, confs, chunksize=1)
-        res.wait(None)
-    except KeyboardInterrupt:
-        print("Caught KeyboardInterrupt, terminating workers")
-        pool.terminate()
+        try:
+            res = pool.map_async(run_config, confs, chunksize=1)
+            res.wait(None)
+            print("Normal termination")
+            pool.close()
+        except KeyboardInterrupt:
+            print("Caught KeyboardInterrupt, terminating workers")
+            pool.terminate()
+
+        pool.join()
     else:
-        print("Normal termination")
-        pool.close()
-    pool.join()
+        for conf in confs:
+            print("Running for {}".format(conf.solver_output_routes_csv))
+            run_config(conf)
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Run Bus Plan.')
+    parser.add_argument('--skip-solver',
+                        dest="skip_solver",
+                        action='store_true',
+                        help='When set, skip running the solver')
+    parser.add_argument('--debug-run',
+                        dest="debug_run",
+                        action='store_true',
+                        help='When set, run a smaller test case')
     parser.add_argument('data_dir',
                         metavar='DATA_DIR',
                         help='Base data directory that holds all properly named input files.')
@@ -149,4 +205,4 @@ if __name__ == "__main__":
                         metavar='CODE_DIR',
                         help='Directory that holds the bus_plan codebase.')
     args = parser.parse_args()
-    run(args.data_dir, args.output_dir, args.code_dir)
+    run(args.data_dir, args.output_dir, args.code_dir, not args.skip_solver, args.debug_run)
